XIVY-15945 remove missing default languages
Previously, if you had a default language set but somehow removed the language from the CMS outside of the CMS Editor (e.g. by deleting the cms yaml file in a file explorer), the default language is still present but cannot be removed in a straightforward way.
Now, whenever the default languages are computed, default languages set in the local storage that are not present in the CMS will be removed.