XIVY-15945 hide missing default languages
Do not remove default languages that are missing from the CMS from the local storage.
Instead, simply filter them out.
The only way to change the default languages is by explicitly checking/unchecking the checkbox in the language tool.
Any other action like adding and deleting languages or opening a CMS with other languages present will not affect the default languages saved in the local storage.