XIVY-15945 handle no default languages in add content object dialog
If no default languages are defined, the "Add Content Object" dialog takes the first language of the CMS to display a field for its value.
This is necessary because it is not possible to create a Content Object without defining at least one value.
This circumstance is conveyed to the user via a message in the "Add Content Object" dialog.