XIVY-15945 fix: update main table on language deletion
When languages are deleted, Content Objects that have no values left because the respective languages have been deleted will be deleted as well.
To make the main table reflect this change, the data query is invalidated if the default languages do not change.
In the case where the default languages do change, this already works, as the table gets rerendered because the columns change.