XIVY-16623 add confirmation dialog when deleting languages with values
If at least one language with values is about to be deleted, clicking the "Save" button will open a confirmation dialog.
The confirmation dialog displays all languages with values about to be deleted with the respective amount of values.
The user can choose to either cancel the deletion or proceed with it.
Everything works using only the keyboard as well.